Use proper `mapexpr` include feature for rewriting included expressions
As of Julia 1.5, `include` has an official feature to rewrite the parsed
expressions - no need manually mess with textual includes and parsing.
Using this feature rather than relying on Base internals is recommended
as it makes the package more robust to future changes in this infrastructure.
In particular, I'm planning to change the module-local `include` to
allow extensions in https://github.com/JuliaLang/julia/pull/55949.
This package is currently doing that, though it could be changed not
to. However, by just using the official feature for what this package
wants to do, we just sidestep the question entirely. The old code is
retained for pre-1.5 compat.